+++ packages/react-core/src/evm/hooks/useSDK.ts
@@ -1,0 +1,39 @@
+import { ThirdwebSDK } from "@thirdweb-dev/sdk";
+import { useContext } from "react";
+import invariant from "tiny-invariant";
+import { TWSDKContext, ThirdwebSDKContext } from "../contexts/thirdweb-sdk";
+
+/**
+ * @internal
+ */
+export function useSDKContext(): TWSDKContext {
+  const ctx = useContext(ThirdwebSDKContext);
+  invariant(
+    ctx._inProvider,
+    "useSDK must be called from within a ThirdwebProvider, did you forget to wrap your app in a <ThirdwebProvider />?",
+  );
+  return ctx;
+}
+
+/**
+ *
+ * @returns {@link ThirdwebSDK}
+ * Access the instance of the thirdweb SDK created by the ThirdwebProvider
+ * to call methods using the connected wallet on the desiredChainId.
+ * @example
+ * ```javascript
+ * const sdk = useSDK();
+ * ```
+ */
+export function useSDK(): ThirdwebSDK | undefined {
+  const { sdk } = useSDKContext();
+  return sdk;
+}
+
+/**
+ * @internal
+ */
+export function useSDKChainId(): number | undefined {
+  const sdk = useSDK();
+  return (sdk as any)?._chainId;
+}