+++ packages/react-core/src/evm/providers/types.ts
@@ -2,7 +2,7 @@ import type { QueryClientProviderProps } from "../../core/providers/query-client
 import type { ThirdwebAuthConfig } from "../contexts/thirdweb-auth";
 import type { Chain } from "@thirdweb-dev/chains";
 import type { SDKOptions } from "@thirdweb-dev/sdk";
-import type { ThirdwebStorage } from "@thirdweb-dev/storage";
+import type { IThirdwebStorage } from "@thirdweb-dev/storage";
 import type { Signer } from "ethers";
 
 export interface ThirdwebSDKProviderProps<TChains extends Chain[]>
@@ -22,7 +22,7 @@ export interface ThirdwebSDKProviderProps<TChains extends Chain[]>
   /**
    * The storage interface to use with the sdk.
    */
-  storageInterface?: ThirdwebStorage;
+  storageInterface?: IThirdwebStorage;
   /**
    * The configuration used for thirdweb auth usage. Enables users to login
    * to backends with their wallet.
@@ -41,8 +41,6 @@ export interface ThirdwebSDKProviderProps<TChains extends Chain[]>
     | (string & {})
     | Chain;
 
-  // api keys that can be passed
-  thirdwebApiKey?: string;
-  alchemyApiKey?: string;
-  infuraApiKey?: string;
+  // client Id for thirdweb services
+  clientId?: string;
 }