+++ packages/react-native/src/evm/utils/global.ts
@@ -1,0 +1,7 @@
+export function isGlobalThisPresent() {
+    return typeof globalThis !== "undefined";
+}
+
+export function isAppBundleIdPresentInGlobal() {
+    return isGlobalThisPresent() && "APP_BUNDLE_ID" in globalThis;
+}