+++ packages/react/src/evm/constants/rpc.ts
@@ -1,2 +1,0 @@
-export const DEFAULT_API_KEY =
-  "f9e23156ccfc6173c5267eaabf320fb692ffab1c2d572095a58091cbb67590cd";