+++ packages/react-native-compat/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @thirdweb-dev/react-native-compat
 
+## 0.2.37
+
+## 0.2.36
+
+### Patch Changes
+
+- [#1316](https://github.com/thirdweb-dev/js/pull/1316) [`d8447146`](https://github.com/thirdweb-dev/js/commit/d8447146092c1962f410155ab2047225453aaa2b) Thanks [@iketw](https://github.com/iketw)! - Adds support for clientId / secretKey
+
+## 0.2.35
+
+## 0.2.34
+
+## 0.2.33
+
 ## 0.2.32
 
 ## 0.2.31